Added security to ensure only a logged in super user can modify and delete products

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -7,6 +8,7 @@ from .models import Product, Category
 from .forms import ModifyProductsForm
 
 
+@login_required()
 def all_products(request):
     """ A view to return all products, including sorting and searching """
     
@@ -69,6 +71,7 @@ def all_products(request):
     return render(request, 'products/products.html', context)
 
 
+@login_required()
 def product_detail(request, product_id):
     """ A view to show individual product details """
 
@@ -81,8 +84,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required()
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        message.error(request, 'This function is only available for superusers')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -102,8 +110,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required()
 def modify_product(request, product_id):
     """ Modify a product in the database """
+    if not request.user.is_superuser:
+        message.error(request, 'This function is only available for superusers')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -127,8 +140,13 @@ def modify_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required()
 def delete_product(request, product_id):
     """ Modify a product in the database """
+    if not request.user.is_superuser:
+        message.error(request, 'This function is only available for superusers')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Item deleted successfully')
